% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveFit1.R
\name{CurveFit1}
\alias{CurveFit1}
\title{This function determines the 4 parameter or 3 parameter log fit for the proteome level curve.}
\usage{
CurveFit1(Data_Quantified)
}
\arguments{
\item{Data_Quantified}{the median abundance values calculated in the Quantify function}
}
\value{
the curve fit parameters for the control and condition curves at the proteome level
}
\description{
This function determines the 4 parameter or 3 parameter log fit for the proteome level curve.
}
\examples{
\dontrun{
Data_CurveFit1Parameters<-CurveFit1(Data_Quantified)
}
}
