% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveFit2.R
\name{CurveFit2}
\alias{CurveFit2}
\title{This function determines the best curve fit for each protein using the data post correction and also determines the R squared for each curve fit}
\usage{
CurveFit2(Data_Corrected)
}
\arguments{
\item{Data_Corrected}{data that meets exclusion criteria from Exclude function}
}
\value{
Curve fits and R squared for each protein
}
\description{
This function determines the best curve fit for each protein using the data post correction and also determines the R squared for each curve fit
}
\examples{
\dontrun{
Data_CurveFit2_Control<-CurveFit2(Data_Corrected_Control)}
}
