% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\title{This function normalizes the abundance values to that measured at the lowest temperature}
\usage{
Normalize(Data_Imported)
}
\arguments{
\item{Data_Imported}{the abundance data imported from Import function}
}
\value{
Normalized data
}
\description{
This function normalizes the abundance values to that measured at the lowest temperature
}
\examples{
\dontrun{
     Data_Normalized<-Normalize(Data_Imported)}
}
