% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDataMelts.R
\name{ReportDataMelts}
\alias{ReportDataMelts}
\title{This function generates results from the Inflect function after applying criteria input from the user}
\usage{
ReportDataMelts(
  Data_Melts,
  Data_CurveFit2_Control,
  Data_CurveFit2_Condition,
  Directory
)
}
\arguments{
\item{Data_Melts}{abundance and fit data for proteins that meet quality criteria in overall workflow}

\item{Data_CurveFit2_Control}{the curve fit data from the Curve Fit 2 function}

\item{Data_CurveFit2_Condition}{the curve fit data from the Curve Fit 2 function}

\item{Directory}{directory where data is saved}
}
\value{
Excel files with summary of data along with melt curve plots for significant proteins
}
\description{
This function generates results from the Inflect function after applying criteria input from the user
}
\examples{
     \dontrun{
     ReportDataMelts(Data_Melts,Data_CurveFit2_Control,Data_CurveFit2_Condition,Directory)}
}
