% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportPANTHER.R
\name{ReportPANTHER}
\alias{ReportPANTHER}
\title{This function generates a PANTHER GO term based plot using the significant melt shifts from analysis}
\usage{
ReportPANTHER(Data_Melts, Directory, Species, PANTHERpvalue, PValMeltFDR)
}
\arguments{
\item{Data_Melts}{abundance and fit data for proteins}

\item{Directory}{directory where results are saved}

\item{Species}{species number for bioinformatics search}

\item{PANTHERpvalue}{pvalue for PANTHER analysis}

\item{PValMeltFDR}{Whether or not the FDR correction for pvalue is used in designation of melts of interest}
}
\value{
Excel files with summary of data along with melt curve plots for significant proteins
}
\description{
This function generates a PANTHER GO term based plot using the significant melt shifts from analysis
}
\examples{
     \dontrun{
     ReportPANTHER(Data_Melts,Directory,Species,PANTHERpvalue,PValMeltFDR)}
}
